Simplify extraction scripts
